add settings utils to use settings outside of components
